requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
pkg-resources==0.0.0
PyYAML==3.13
requests==2.19.1
urllib3==1.23
